+++ Changelog.md
@@ -1,5 +1,10 @@
 # Gifify Changelog
 
+## 2013-03-20
+
+* Upload to CloudApp by default (prevent with `-n` flag)
+* Passing `-x` flag removes resulting .gif *and* original file.
+
 ## 2013-03-14
 
 * Initial release